LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/columns-shorthand-reset-wrap-expected.txt


FAIL multicol1 assert_equals: expected (string) "auto" but got (undefined) undefined
FAIL multicol2 assert_equals: expected (string) "auto" but got (undefined) undefined
FAIL multicol3 assert_equals: expected (string) "nowrap" but got (undefined) undefined
